requirements.txt
neo4j>=5.20.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
google-generativeai>=0.7.2
python-dotenv>=1.0.1
pydantic>=2.8.2
httpx>=0.27.0
rich>=13.7.1

